% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{bcspecies-package}
\alias{bcspecies}
\alias{bcspecies-package}
\title{bcspecies: Visualize and Analyze BC Species Conservation Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

For a given species in BC, map presence and absence by ecosection and view BC List conservation status and COSEWIC status.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GinKampen/bcspecies}
  \item Report bugs at \url{https://github.com/GinKampen/bcspecies}
}

}
\author{
\strong{Maintainer}: Virginia Kampen \email{virginiakampen@gmail.com}

Authors:
\itemize{
  \item Sebastian Dalgarno \email{seb@poissonconsulting.ca} [contributor]
}

}
\keyword{internal}
