% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{conservation_status}
\alias{conservation_status}
\title{conservation_status}
\usage{
conservation_status(species)
}
\arguments{
\item{species}{A string of the species.}
}
\value{
A data.frame.
}
\description{
Get BCList, COSEWIC Status and Implemented Date of COSEWIC status for a valid species.
}
\examples{
conservation_status("Anemone occidentalis - Carex nigricans")
 

}
