% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_map}
\alias{species_map}
\title{species_map}
\usage{
species_map(species)
}
\arguments{
\item{species}{A string of the species.}
}
\value{
A ggplot object.
}
\description{
Map distribution of species by BC ecosection.
}
\examples{
species_map("Anemone occidentalis - Carex nigricans")

}
