% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{bcspecies-package}
\alias{bcspecies}
\alias{bcspecies-package}
\title{bcspecies: viewing and working with clean dataset of British Columbia species}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The clean dataset of B.C. species allows the user to use two functions; one that provides an interactive map of the specie's ecosection, and the other which provides the user with three components of the conservation status.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/pkgtemplate}
  \item Report bugs at \url{https://github.com/poissonconsulting/pkgtemplate/issues}
}

}
\author{
\strong{Maintainer}: Virginia Kampen \email{virginiakampen@gmail.com}

Authors:
\itemize{
  \item Sebastian Dalgarno \email{seb@poissonconsulting.ca} [contributor]
}

}
\keyword{internal}
