% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{conservation_status}
\alias{conservation_status}
\title{conservation_status}
\usage{
conservation_status(species)
}
\arguments{
\item{species}{the species provided}
}
\value{
A table containing results of BcList, COSWEIC status and Implemented COSWEIC date
}
\description{
Allows user to view table containing results of BCList, COSEWIC Status and Implemented COSEWIC date
}
\examples{
conservation_status("Anemone occidentalis - Carex nigricans")
 

}
