% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_map}
\alias{species_map}
\title{species_map}
\usage{
species_map(species)
}
\arguments{
\item{species}{the species provided}
}
\value{
a map containing highlighted ecosections connected to species provided
}
\description{
Allows user to view map containing highlighted ecosections connected to species provided
}
